Remove ipdb
I removed error handling altogether. There wasn't any in xor's
previous versions of the code and I think at this point if the
script fails to tweet we want it to fail loudly.

At some point we'll want to integrate logging and record failures
while allowing the script to continue