Use `requests.Session` for HTTP requests
This will improve performance when making multiple requests to the
same domain or URL and is generally recommended when using the
`requests` library for a lot of requests.

The change does require passing the session object around a little
bit, and I think this does expose some places where the `Article`
class is doing a little too much, but I haven't done anything super
complicated here except to move `canonicalize_url` out of the
`__init__` method.  More work can be done here in the future.

I've also standardized the `timeout` argument for requests into a
global constant, but it might be something we want to put into the
config file or the command line arguments.

Furthermore, I've changed our use of `feedparser` to no longer request
the feed, only to parse the contents that we've fetched with our
requests session, which allows us to control the timeout behavior
there (feedparser does not expose this option).